Move database setup to conftest.py
